python/requirements.txt
requests>=2.20.0
PyYAML>=4.2b1
scikit-learn
matplotlib